+++ client/consul_discovery.go
@@ -28,6 +28,8 @@ type ConsulDiscovery struct {
 	// -1 means it always retry to watch until zookeeper is ok, 0 means no retry.
 	RetriesAfterWatchFailed int
 
+	filter ServiceDiscoveryFilter
+
 	stopCh chan struct{}
 }
 
@@ -65,7 +67,11 @@ func NewConsulDiscoveryStore(basePath string, kv store.Store) ServiceDiscovery {
 	prefix := d.basePath + "/"
 	for _, p := range ps {
 		k := strings.TrimPrefix(p.Key, prefix)
-		pairs = append(pairs, &KVPair{Key: k, Value: string(p.Value)})
+		pair := &KVPair{Key: k, Value: string(p.Value)}
+		if d.filter != nil && !d.filter(pair) {
+			continue
+		}
+		pairs = append(pairs, pair)
 	}
 	d.pairs = pairs
 	d.RetriesAfterWatchFailed = -1
@@ -95,6 +101,11 @@ func NewConsulDiscoveryTemplate(basePath string, consulAddr []string, options *s
 // Clone clones this ServiceDiscovery with new servicePath.
 func (d ConsulDiscovery) Clone(servicePath string) ServiceDiscovery {
 	return NewConsulDiscoveryStore(d.basePath+"/"+servicePath, d.kv)
+}
+
+// SetFilter sets the filer.
+func (d ConsulDiscovery) SetFilter(filter ServiceDiscoveryFilter) {
+	d.filter = filter
 }
 
 // GetServices returns the servers
@@ -172,7 +183,11 @@ func (d *ConsulDiscovery) watch() {
 				var pairs []*KVPair // latest servers
 				for _, p := range ps {
 					k := strings.TrimPrefix(p.Key, prefix)
-					pairs = append(pairs, &KVPair{Key: k, Value: string(p.Value)})
+					pair := &KVPair{Key: k, Value: string(p.Value)}
+					if d.filter != nil && !d.filter(pair) {
+						continue
+					}
+					pairs = append(pairs, pair)
 				}
 				d.pairs = pairs