+++ client/etcd_discovery.go
@@ -29,6 +29,8 @@ type EtcdDiscovery struct {
 	// -1 means it always retry to watch until zookeeper is ok, 0 means no retry.
 	RetriesAfterWatchFailed int
 
+	filter ServiceDiscoveryFilter
+
 	stopCh chan struct{}
 }
 
@@ -76,7 +78,11 @@ func NewEtcdDiscoveryStore(basePath string, kv store.Store) ServiceDiscovery {
 			}
 		}
 		k := strings.TrimPrefix(p.Key, prefix)
-		pairs = append(pairs, &KVPair{Key: k, Value: string(p.Value)})
+		pair := &KVPair{Key: k, Value: string(p.Value)}
+		if d.filter != nil && !d.filter(pair) {
+			continue
+		}
+		pairs = append(pairs, pair)
 	}
 	d.pairs = pairs
 	d.RetriesAfterWatchFailed = -1
@@ -103,6 +109,11 @@ func NewEtcdDiscoveryTemplate(basePath string, etcdAddr []string, options *store
 // Clone clones this ServiceDiscovery with new servicePath.
 func (d EtcdDiscovery) Clone(servicePath string) ServiceDiscovery {
 	return NewEtcdDiscoveryStore(d.basePath+"/"+servicePath, d.kv)
+}
+
+// SetFilter sets the filer.
+func (d EtcdDiscovery) SetFilter(filter ServiceDiscoveryFilter) {
+	d.filter = filter
 }
 
 // GetServices returns the servers
@@ -195,7 +206,11 @@ func (d *EtcdDiscovery) watch() {
 						}
 					}
 					k := strings.TrimPrefix(p.Key, prefix)
-					pairs = append(pairs, &KVPair{Key: k, Value: string(p.Value)})
+					pair := &KVPair{Key: k, Value: string(p.Value)}
+					if d.filter != nil && !d.filter(pair) {
+						continue
+					}
+					pairs = append(pairs, pair)
 				}
 				d.pairs = pairs