+++ client/etcdv3_discovery.go
@@ -29,6 +29,8 @@ type EtcdV3Discovery struct {
 	// -1 means it always retry to watch until zookeeper is ok, 0 means no retry.
 	RetriesAfterWatchFailed int
 
+	filter ServiceDiscoveryFilter
+
 	stopCh chan struct{}
 }
 
@@ -79,7 +81,11 @@ func NewEtcdV3DiscoveryStore(basePath string, kv store.Store) ServiceDiscovery {
 			continue
 		}
 		k := strings.TrimPrefix(p.Key, prefix)
-		pairs = append(pairs, &KVPair{Key: k, Value: string(p.Value)})
+		pair := &KVPair{Key: k, Value: string(p.Value)}
+		if d.filter != nil && !d.filter(pair) {
+			continue
+		}
+		pairs = append(pairs, pair)
 	}
 	d.pairs = pairs
 	d.RetriesAfterWatchFailed = -1
@@ -106,6 +112,11 @@ func NewEtcdV3DiscoveryTemplate(basePath string, etcdAddr []string, options *sto
 // Clone clones this ServiceDiscovery with new servicePath.
 func (d EtcdV3Discovery) Clone(servicePath string) ServiceDiscovery {
 	return NewEtcdV3DiscoveryStore(d.basePath+"/"+servicePath, d.kv)
+}
+
+// SetFilter sets the filer.
+func (d EtcdV3Discovery) SetFilter(filter ServiceDiscoveryFilter) {
+	d.filter = filter
 }
 
 // GetServices returns the servers
@@ -202,7 +213,11 @@ func (d *EtcdV3Discovery) watch() {
 					}
 
 					k := strings.TrimPrefix(p.Key, prefix)
-					pairs = append(pairs, &KVPair{Key: k, Value: string(p.Value)})
+					pair := &KVPair{Key: k, Value: string(p.Value)}
+					if d.filter != nil && !d.filter(pair) {
+						continue
+					}
+					pairs = append(pairs, pair)
 				}
 				d.pairs = pairs