+++ client/xclient.go
@@ -58,12 +58,18 @@ type KVPair struct {
 	Value string
 }
 
+// ServiceDiscoveryFilter can be used to filter services with customized logics.
+// Servers can register its services but clients can use the customized filter to select some services.
+// It returns true if ServiceDiscovery wants to use this service, otherwise it returns false.
+type ServiceDiscoveryFilter func(kvp *KVPair) bool
+
 // ServiceDiscovery defines ServiceDiscovery of zookeeper, etcd and consul
 type ServiceDiscovery interface {
 	GetServices() []*KVPair
 	WatchService() chan []*KVPair
 	RemoveWatcher(ch chan []*KVPair)
 	Clone(servicePath string) ServiceDiscovery
+	SetFilter(ServiceDiscoveryFilter)
 	Close()
 }